OJ-2532: Move resources folder
Move the resources folder in stepfunction/aws
to be a sibling of the api-gateway and step-function
folder as it contains resources referenced by
both api-gateway and step-function
