+++ integration-tests/resources/aws-helper.ts
