+++ integration-tests/resources/cloudformation-helper.ts
