+++ integration-tests/resources/dynamodb-helper.ts
