+++ integration-tests/resources/kms-helper.ts
