+++ integration-tests/resources/ssm-param-helper.ts
