+++ integration-tests/resources/stepfunction-helper.ts
