+++ integration-tests/step-functions/aws/check-session/check-session.test.ts
@@ -1,6 +1,6 @@
-import { stackOutputs } from "../resources/cloudformation-helper";
-import { clearItems, populateTable } from "../resources/dynamodb-helper";
-import { executeStepFunction } from "../resources/stepfunction-helper";
+import { stackOutputs } from "../../../resources/cloudformation-helper";
+import { clearItems, populateTable } from "../../../resources/dynamodb-helper";
+import { executeStepFunction } from "../../../resources/stepfunction-helper";
 
 describe("check-session", () => {
   const input = {
